requirements.txt
pyautogen>=0.2.0
python-dotenv==1.0.0
Pillow==10.0.0
requests==2.31.0
openai>=1.0.0
gTTS>=2.3.1
